frontend/requirements.txt
streamlit
requests
pandas
streamlit
python-dotenv
reportlab